simplify the template repo
