+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
-    "python.analysis.extraPaths": [
-        "./lib"
-    ],
-    "python.envFile": "${workspaceFolder}/.env"
+    "python.envFile": "${workspaceFolder}/venv",
+    "python.testing.pytestArgs": [],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true
 }
