--- cookiecutter.json
@@ -1,5 +1,0 @@
-{
-    "full_name": "Amiga User",
-    "app_name": "Hello World",
-    "package_name": "{{cookiecutter.app_name.replace(' ', '_').replace('-', '_')}}"
-}
