+++ setup.cfg
@@ -1,0 +1,25 @@
+# NOTE: update the `[project]` section with your app details
+[project]
+name = amiga_pkg
+version = 0.0.1
+author = Farm-ng, Inc.
+author_email = contact@farm-ng.com
+description = Template app for the Amiga robot
+
+# NOTE: update the the requirements for your app
+[options]
+python_requires = >=3.6
+setup_requires =
+    wheel
+install_requires =
+    wheel
+    kivy
+    farm_ng_amiga
+
+# NOTE: do not touch this path unless you know what you are doing
+package_dir =
+    = libs
+
+[options.extras_require]
+dev =
+    pytest
