+++ setup.cfg.md5sum
@@ -1,0 +1,1 @@
+d052a9e3eecaa3d3f98d745fb6ed29c2  /data/home/amiga/amiga-app-template/setup.cfg
