+++ src/res/main.kv
@@ -1,0 +1,16 @@
+RelativeLayout:
+    Button:
+        id: back_btn_layout
+        pos_hint: {"x": 0.0, "top": 1.0}
+        background_color: 0, 0, 0, 0
+        size_hint: 0.1, 0.1
+        background_normal: "assets/back_button.png"
+        on_release: app.on_exit_btn()
+        Image:
+            source: "assets/back_button_normal.png" if self.parent.state == "normal" else "assets/back_button_down.png"
+            pos: self.parent.pos
+            size: self.parent.size
+    Label:
+        id: counter_label
+        text: "Tic: 0"
+        font_size: 40
