--- {{cookiecutter.package_name}}/pyproject.toml
@@ -1,6 +1,0 @@
-[build-system]
-requires = [
-    "setuptools",
-    "wheel",
-]
-build-backend = "setuptools.build_meta"
