--- {{cookiecutter.package_name}}/setup.cfg
@@ -1,18 +1,0 @@
-[metadata]
-name = {{ cookiecutter.package_name }}
-
-
-[options]
-setup_requires =
-    wheel
-install_requires =
-    wheel
-    kivy
-    farm_ng_amiga
-package_dir =
-    = apps
-packages = find:
-python_requires = >=3.6
-
-[options.packages.find]
-where = apps
